+++ src/test/java/pl/marczyk/parkinglotapi/service/cost/FiveMinuteCostRuleTest.java
@@ -3,6 +3,8 @@ package pl.marczyk.parkinglotapi.service.cost;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
+import java.math.BigDecimal;
+
 import static org.junit.jupiter.api.Assertions.*;
 
 class FiveMinuteCostRuleTest {
@@ -16,7 +18,7 @@ class FiveMinuteCostRuleTest {
             "6, 1",
             "10, 2"
     })
-    void should_apply_one_pound_for_every_five_minute_parking(int minutes, double expected) {
+    void should_apply_one_pound_for_every_five_minute_parking(int minutes, BigDecimal expected) {
         assertEquals(expected, cut.apply(minutes));
     }
 }
